requirements.txt
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
awscli==1.18.69
background==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
botocore==1.16.19
Brlapi==0.6.6
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==6.7
cloud-init==21.1
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.3.1
cupshelpers==1.0
defer==1.0.6
defusedxml==0.7.1
distro-info===0.18ubuntu0.18.04.1
Django==3.2.5
django-allauth==0.45.0
django-background-tasks==1.2.5
django-bootstrap4==2.3.1
django-compat==1.0.15
django-cors-headers==3.7.0
django-model-utils==4.1.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
docutils==0.14
feedparser==5.2.1
html5lib==0.999999999
httplib2==0.11.3
hyperlink==17.3.1
idna==2.10
importlib-metadata==2.0.0
incremental==16.10.1
Jinja2==2.10
jmespath==0.9.3
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
kazam==1.5.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==2.0.3
netifaces==0.10.4
oauth==1.0.1
oauthlib==3.1.1
olefile==0.45.1
onboard==1.4.1
packaging==21.0
PAM==0.4.2
paramiko==2.7.2
pexpect==4.2.1
Pillow==5.1.0
prettytable==0.7.2
protobuf==3.0.0
psutil==5.8.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.2.0
pygobject==3.26.1
PyJWT==2.1.0
pymacaroons==0.13.0
PyMySQL==0.8.0
PyNaCl==1.4.0
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyRFC3339==1.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
roman==2.0.0
rsa==3.4.2
s3transfer==0.3.3
SecretStorage==2.3.1
service-identity==16.0.0
simplejson==3.13.2
six==1.16.0
sos==4.1
soupsieve==2.0.1
sqlparse==0.4.1
ssh-import-id==5.7
system-service==0.3
systemd-python==234
Twisted==17.9.0
typing-extensions==3.10.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.26.3
usb-creator==0.3.3
wadllib==1.3.2
webencodings==0.5
xdiagnose==3.8.8
xkit==0.0.0
zipp==3.3.1
zope.interface==4.3.2
